my-app/backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
firebase-admin==6.5.0
Requests==2.31.0
